Use the batched queue implementation, and remove the non-batched option
